requirements.txt
Django
torch
torchvision
pillow
numpy
scikit-learn
matplotlib
seaborn
opencv-python